# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Sistema de Inventario",
        "description": "Documentación oficial de la API para la gestión de inventarios, áreas y bienes.",
        "contact": {
            "email": "soporte@tuempresa.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8081/api",
            "description": "Servidor Principal de la API"
        }
    ],
    "paths": {
...
        "/formularios/roles": {
            "get": {
                "tags": [
                    "Formularios"
                ],
                "summary": "Obtener lista completa de roles",
                "description": "Obtener lista de roles\n\nDevuelve una lista de todos los roles disponibles en el sistema.\nSe utiliza para poblar selectores en formularios de creación/edición de usuarios.",
                "operationId": "getRolFormOptions",
                "responses": {
                    "200": {
                        "description": "Lista de roles obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "rol_nombre": {
                                                "type": "string",
                                                "example": "Administrador"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            }
        },
...
    },
...
}